add downloading from pinata

+++ src/helpers/pinata.py
@@ -4,24 +4,39 @@ from rosbag import Bag
 import requests
 import rospy
 
-def pinata_download(ipfs_hash: str, mode: str = "r") -> str:
+def pinata_rosbag(ipfs_hash: str) -> str:
 
-    try:
-        data = requests.get(f'https://gateway.pinata.cloud/ipfs/{ipfs_hash}')
-    except Exception as e:
-        rospy.loginfo(f'downloading error: {e}')
-    
-    tmpfile = NamedTemporaryFile(delete=False)
-    tmpfile.write(data.content)
-    bag = Bag(tmpfile.name)
-    topics_list = bag.get_type_and_topic_info()[1].keys()
-    messages = {}
+    data = requests.get(f'https://gateway.pinata.cloud/ipfs/{ipfs_hash}')
+    rospy.loginfo(data.status_code)
+    if data.status_code != 200:
+        rospy.loginfo(f'Error downloading rosbag! {data.content}')
+        return
 
-    for topic, msg, timestamp in bag.read_messages():
-        if topic not in messages:
-            messages[topic] = [msg]
-        else:
-            messages[topic].append(msg)
+    else:
+        tmpfile = NamedTemporaryFile(delete=False)
+        tmpfile.write(data.content)
+        bag = Bag(tmpfile.name)
+        
+        messages = {}
+        for topic, msg, timestamp in bag.read_messages():
+            if topic not in messages:
+                messages[topic] = [msg]
+            else:
+                messages[topic].append(msg)
 
-    os.unlink(tmpfile.name)
+        os.unlink(tmpfile.name)
     return (messages, bag)
+
+
+def pinata_download(hash: str, mode: str = "r") -> str:
+    
+    res = requests.get(f'https://gateway.pinata.cloud/ipfs/{hash}')
+
+    if res.status_code != 200:
+        rospy.loginfo(f'Error while downloading from Pinata: {res.content}')
+
+    else:
+        tmpfile = NamedTemporaryFile(delete=False)
+        tmpfile.write(res.content)
+        with open(tmpfile.name, mode) as f:
+            return f.read()
